Call_tree inheritance refactored:
Nodes implementation for unordered and ordered call tree separated.
Updater and concurrent tree now work only with ordered call tree.
This is required because fields in ordered and unordered trees are
different and we need to understand how to update them.
merge_into and to_json fields are virtual and overloaded in each tree.